+++ src/rules/namespace.js
@@ -2,7 +2,30 @@ import Exports from '../ExportMap'
 import importDeclaration from '../importDeclaration'
 import declaredScope from 'eslint-module-utils/declaredScope'
 
-module.exports = function (context) {
+exports.meta = {
+  schema: [
+    {
+      'type': 'object',
+      'properties': {
+        'allowComputed': {
+          'description':
+            'If `false`, will report computed (and thus, un-lintable) references ' +
+            'to namespace members.',
+          'type': 'boolean',
+          'default': false,
+        },
+      },
+      'additionalProperties': false,
+    },
+  ],
+}
+
+exports.create = function namespaceRule(context) {
+
+  // read options
+  const {
+    allowComputed = false,
+  } = context.options[0] || {}
 
   const namespaces = new Map()
 
@@ -91,9 +114,11 @@ module.exports = function (context) {
              dereference.type === 'MemberExpression') {
 
         if (dereference.computed) {
-          context.report(dereference.property,
-            'Unable to validate computed reference to imported namespace \'' +
-            dereference.object.name + '\'.')
+          if (!allowComputed) {
+            context.report(dereference.property,
+              'Unable to validate computed reference to imported namespace \'' +
+              dereference.object.name + '\'.')
+          }
           return
         }
 
@@ -104,9 +129,12 @@ module.exports = function (context) {
           break
         }
 
+        const exported = namespace.get(dereference.property.name)
+        if (exported == null) return
+
         // stash and pop
         namepath.push(dereference.property.name)
-        namespace = namespace.get(dereference.property.name).namespace
+        namespace = exported.namespace
         dereference = dereference.parent
       }