+++ tests/files/typescript.ts
@@ -1,0 +1,5 @@
+type X = string
+
+export function getFoo() : X {
+  return "foo"
+}