+++ tests/src/rules/no-nodejs-modules.js
@@ -5,10 +5,10 @@ import { RuleTester } from 'eslint'
 const ruleTester = new RuleTester()
     , rule = require('rules/no-nodejs-modules')
 
-const errors = [{
+const error = message => ({
   ruleId: 'no-nodejs-modules',
-  message: 'Do not import Node.js builtin modules',
-}]
+  message,
+})
 
 ruleTester.run('no-nodejs-modules', rule, {
   valid: [
@@ -26,23 +26,60 @@ ruleTester.run('no-nodejs-modules', rule, {
     test({ code: 'var foo = require("foo")'}),
     test({ code: 'var foo = require("./")'}),
     test({ code: 'var foo = require("@scope/foo")'}),
+    test({
+      code: 'import events from "events"',
+      options: [{
+        allow: ['events'],
+      }],
+    }),
+    test({
+      code: 'import path from "path"',
+      options: [{
+        allow: ['path'],
+      }],
+    }),
+    test({
+      code: 'var events = require("events")',
+      options: [{
+        allow: ['events'],
+      }],
+    }),
+    test({
+      code: 'var path = require("path")',
+      options: [{
+        allow: ['path'],
+      }],
+    }),
+    test({
+      code: 'import path from "path";import events from "events"',
+      options: [{
+        allow: ['path', 'events'],
+      }],
+    }),
   ],
   invalid: [
     test({
       code: 'import path from "path"',
-      errors,
+      errors: [error('Do not import Node.js builtin module "path"')],
     }),
     test({
       code: 'import fs from "fs"',
-      errors,
+      errors: [error('Do not import Node.js builtin module "fs"')],
     }),
     test({
       code: 'var path = require("path")',
-      errors,
+      errors: [error('Do not import Node.js builtin module "path"')],
     }),
     test({
       code: 'var fs = require("fs")',
-      errors,
+      errors: [error('Do not import Node.js builtin module "fs"')],
+    }),
+    test({
+      code: 'import fs from "fs"',
+      options: [{
+        allow: ['path'],
+      }],
+      errors: [error('Do not import Node.js builtin module "fs"')],
     }),
   ],
 })