+++ tests/src/utils.js
@@ -57,7 +57,7 @@ export const SYNTAX_CASES = [
   test({ code: 'export default x' }),
   test({ code: 'export default class x {}' }),
 
-  // issue #267: parser whitelist
+  // issue #267: parser opt-in extension list
   test({
     code: 'import json from "./data.json"',
     settings: { 'import/extensions': ['.js'] }, // breaking: remove for v2
@@ -83,5 +83,9 @@ export const SYNTAX_CASES = [
   test({
     code: 'export * from "./issue-370-commonjs-namespace/bar"',
     settings: { 'import/ignore': ['foo'] },
+  }),
+
+  test({
+    code: 'import * as a from "./commonjs-namespace/a"; a.b',
   }),
  ]