+++ utils/ignore.js
@@ -13,13 +13,28 @@ function validExtensions(context) {
   }
 
   lastSettings = context.settings
-  cachedSet = new Set(context.settings['import/extensions'] || [ '.js' ])
+  cachedSet = makeValidExtensionSet(context.settings)
   return cachedSet
+}
+
+function makeValidExtensionSet(settings) {
+  // start with explicit JS-parsed extensions
+  const exts = new Set(settings['import/extensions'] || [ '.js' ])
+
+  // all alternate parser extensions are also valid
+  if ('import/parsers' in settings) {
+    for (let parser in settings['import/parsers']) {
+      settings['import/parsers'][parser]
+        .forEach(ext => exts.add(ext))
+    }
+  }
+
+  return exts
 }
 
 exports.default = function ignore(path, context) {
   // check extension whitelist first (cheap)
-  if (!validExtensions(context).has(extname(path))) return true
+  if (!hasValidExtension(path, context)) return true
 
   if (!('import/ignore' in context.settings)) return false
   const ignoreStrings = context.settings['import/ignore']
@@ -34,3 +49,8 @@ exports.default = function ignore(path, context) {
 
   return false
 }
+
+function hasValidExtension(path, context) {
+  return validExtensions(context).has(extname(path))
+}
+exports.hasValidExtension = hasValidExtension